requirements.txt
torch
numpy
tqdm
wandb
PyYAML
soundfile
librosa
datasets
sentencepiece
matplotlib
jiwer
torchcodec
hydra-core
omegaconf
